+++ test_deps.ts
@@ -2,6 +2,9 @@ export { assertType } from "https://deno.land/std@0.187.0/testing/types.ts";
 export type { IsExact } from "https://deno.land/std@0.187.0/testing/types.ts";
 
 export {
+    assert,
+    assertEquals,
     assertInstanceOf,
     assertStrictEquals,
+    assertThrows,
 } from "https://deno.land/std@0.187.0/testing/asserts.ts";